Delete unnecessary dtos
Removed unnecessary DTOs for LogEntry. The LogEntry will be created on the server upon receiving an HTTP request. Therefore, data specific to LogEntry will not be passed through the request body. User information will be extracted from the JWT token, which is required for successful requests to the LogEntry endpoints.